[OPTIONAL] refactor(config): Move git user config to its own `user` section in the config
We are conceptually manipulating a "generic" username rather than something inherently attached to `git` - `git` is simply the default way by which this username is deduced. Thus we configure it as its own object, using `git` data as a default, and show a warning if the two are ever out of sync.