feat(git): Make `Git.get_working_tree_changes` use a temporary index instead of manually diffing every file
This is much higher performance, as only a fixed number of calls to `git` have to be made (considering the speed of git vs Python itself, we can probably say this makes the function O(1) instead of O(n)).
Moreover, this makes the function work properly on Windows, as we do not require `/dev/null` to exist anymore.