feat(build): Add `build` module and `BuildMetadata` class + tests
The metadata schema is based on the "Build UX" section of the "Future of Agent development" document from @ofek
These `BuildMetadata` objects are meant to be serialized to a file and dropped along with the build artifact.
Therefore it is important to make sure that they are frozen and can be nicely serialized to a file.