refactor(git): Make `Commit` objects frozen
We can actually do this quite easily by avoiding the use of `_details` and `_changes` fields,
and instead using `__dict__` to store the details and changes.
This is the same technique that `cached_property` uses internally,
so trying to access the cached property works as expected if the details or changes have been set by one of the `get_details_from_*` or `get_changes_from_*` methods.