Fix erroneous results for directories when using `dda info owners code` (#202)
Spurred on by this Slack thread: https://dd.slack.com/archives/C04U1B77E84/p1759782594188839

Problem description
The codeowners library imparts meaning on trailing slashes in the arguments we feed it: a path is considered to be a directory if and only if it ends in a trailing slash.

When passing args on the CLI, they get converted to pathlib.Path objects, which always strip a trailing slash. See this issue

Moreover, the fact that a trailing slash is required when querying directory ownership information is quite unintuitive and makes for a bad DevX: for example, some shells won't add a trailing slash when tabbing.

Proposed solution
To get around this, we rely on the fact that we enforce the existence of the paths we are querying owners for. Since they must exist, we can just have the command check whether it is a directory or not, and if so ensure a trailing slash is present.

Extra changes
The last commits in this PR also refactor the unit tests defined for dda info owners code:

We pre-create test file structures in a fixtures/ directory instead of creating them on run. This allows us to make clearer the example file structures we are testing against, and also avoids the need for file creation fixtures.
The performance of the tests is improved by skipping the dependency synchronization step that would otherwise run on every call, and which is responsible for most of the test's runtime. e096ad8cf6374382aba0085bdb6e24d2a62d9c2b