[ACIX-1225] Create `-v` option on dev env start command (#230)
* feat(env): Add ability to mount arbitrary volumes into linux container dev envs

* feat(test): Add test for creating dev env with extra bind mounts

* refactor(env): Use `Mount` class for specifying extra mounts

* feat(env): Add validation logic for `-v` parameter

* fix(tests): Fix test after using `Mount` class and adding validation logic

* feat(tests): Add test for invalid `-v` params

* fix(env): Fix split on colons

* fix(env): Disallow passing literal `~` in paths

This should be interpolated by the shell before it ever gets to Python anyway.

* feat(env): Only allow bind mounts with `-v`

For more complex usecases like using named volumes or weird flags, users should use `--mount`, which will be implemented next.

* feat(env): Add support for `--mount` parameter

* fix(tests): Skip `--mount` invalid param tests on Windows

They would be annoying to fix, and there should not be any Windows-specific behavior here.
The positive test cases are still run on Windows.

* refactor(env): Shorten usage examples for brevity

* fix(env): Remove Windows-incompatible processing logic for `-v`

We just pass the arguments straight to `docker run`, with no extra validation, instead. This means we also have to duplicate the test to test separately for `-v` and `-m` options.

* refactor: Pass `--mount` arguments directly to `docker run` + remove useless validation logic

* Final help text cleanup

Co-authored-by: Ofek Lev <ofekmeister@gmail.com>

---------

Co-authored-by: Ofek Lev <ofekmeister@gmail.com>